Fixed parcel bugs in Account and Label entities; TempDaoImpl insertWithRetry recycle parcel fix

+++ app/src/main/java/me/gingerninja/authenticator/data/db/entity/AbstractAccount.java
@@ -104,7 +104,7 @@ abstract class AbstractAccount {
     public static void restoreFromParcel(@NonNull Account account, @NonNull Parcel parcel) {
         parcel.setDataPosition(0);
 
-        account.id = parcel.readInt();
+        account.id = parcel.readLong();
         account.setUid(parcel.readString());
         account.setTitle(parcel.readString());
         account.setType(parcel.readString());
